Fixed bug where send behavior did not work
The previous implementation worked for SwiftKey and Fleksy but not the
AOSP or Google keyboards. Those did not send an onEditorAction when
SwiftKey and Fleksy did. By removing the TYPE_TEXT_FLAG_MULTI_LINE bit
we can force all keyboards to assume that adding a new line would be a
bad idea. Combined with a "send" IME option already set in the XML by
Hangouts this should display a special send button depending on the
specific keyboard.

Thanks to XDA member ripv2 for reporting.