Customize MMS compression level, format, and scale
Previously, XHangouts sent all MMS as PNGs. Nearly all MMSCs should
accept these, but PNG compression isn't always going to be the best
choice. Verizon, for instance, converts your PNG input to a ~30 KB
JPEG. It's now possible to select PNG or JPEG compression, which
Android natively supports, along with the desired quality level if JPEG
is chosen. The default has been set to JPEG at Q60, a very reasonable
size-compression-quality balance.

Perhaps just as useful is the ability to customize the exact resolution
you wish to target. 640x640 is the default. XHangouts now ignores the
"suggested" resolution from GHangouts and this should fix related bugs.

These new options are included in two neat little custom AlertDialogs.
The resize/scale dialog includes some additional information about
the target resolution. The image type and quality dialog is mighty
impressive. It includes a spinner for the format and a live seekbar for
the quality. But what's cool is that it has two example images that
will demo your current quality options. The first is a colorful high
complexity photo and the second is a screenshot with hard edges and
solid colors. To avoid licensing issues, the photo is my own and is
suitable for use in XHangouts and forks. Overlaid on each image is
text relaying the size in kilobytes of the resulting image. The exact
size is not important (and will change depending on your device's PPI)
but rather its size relative to other quality levels.

I used a Q100 JPEG for the photo because it would have significantly
increased the size of the APK if I went with a PNG. Since the photo is
displayed on such a small portion of the screen (about 300dp by 150dp)
no user will notice a difference in quality and is suitable for
comparison. If the need arises to re-encode either of the two images
the original crops are stored in PNG format in the repo.

On a different note, the full scratchspace image URI is no longer sent
directly to the debug log but is truncated after the query string. This
prevents someone's personal email from being shown if they post their
debug logs.