Introducing MMS APN Splicing
This commit represents over a week worth of reverse engineering,
research, and testing.

MMS on Android is problematic for reasons which I'll refrain from
detailing here. MMS APN Splicing is a feature which will cut out the
standard calls from Hangouts to startUsingNetworkFeature and accelerate
the MMS transaction process. This should also alleviate msm_hsic_host
related kernel wakelocks as the modem is no longer being requested to
turn an interface on and off nearly every transaction. My device, while
in the midst of that wakelock, among others, achieved a toasty 110.3 °F
(43.5 °C). Hot enough that holding your finger on the glass for more
than a few seconds would be very uncomfortable. Battery usage would
rival roughly one percent per minute.

Some refactoring still needs to be done but that applies to the module
as a whole. Once this implementation matures it should be capable of
receiving MMS through WiFi and gracefully falling back to the standard
Hangouts APN connectivity logic when a mobile connection isn't present.
Configuration via the SettingsActivity still needs to be added. Reading
the default MMS APN information in automatically would be ideal but
allowing users to manually specify MMSC connection info is also
a good idea.

The current known limitations of this implementation are as follows:
 - Verizon's MMSC is hardcoded in, it will only work as-is for VZW
 - It will probably completely fail when on a WiFi connection
 - If no network is available it will fail
 - Graceful degradation is planned but not supported
 - Non-configurable. It turns on and off with the module.
 - Lots of nonsensical debug messages :)

The documentation created alongside the research for this feature is
available here: http://goo.gl/Xfa13j