Add ability to hide the video call button
Suggested by PhaseBurn @ XDA. This will remove the video call button
from the ActionBar/Toolbar of a Hangouts conversation. It will continue
to be accessible in the overflow menu.

Limitations to the current approach are many and must be redone before
it can be considered widely useful. It will definitely break if the
Hangouts app is not using the English lang or if the Video call text
ever changes (in any language). It's not very performant either as it
must loop through the list of available menu items. This may not be
fixable but it would be nice to obtain a directly accessible ID. Loop
through once, find the offset, store it in memory, and hope it never
changes?

The best way I can think of to go about fixing this is to read the
resources.arsc file of Hangouts and locate the MenuItem XML definition
ID. Since Xposed does not provide an API for this it is far less
trivial than it seems. The Android aapt tool is capable of doing this
so looking into the source for that tool should yield a way to do it in
XHangouts.