Properly colorize tabs and progress indicators
Brought to you by a ColorUtils class I've compiled and adapted from
various sources noted in the comments.

I wanted to avoid having any source files from Hangouts included in
XHangouts due to possible copyright issues. Opening up the 9-patches in
GIMP and creating different one for each style was off the table. The
only other way to style the tab and (to a lesser degree) progress
indicator drawables is to do it in-place. I experimented with copying
the bitmaps into memory and then manually changing the hue but turns out
it's possible to do this much more easily with a Android ColorFilter +
ColorMatrix. The only downside is that ColorFilters will change the hue
by a certain offset instead of an absolute position. The workaround was
to calculate the difference between the original and the desired colors.
That offset would be used to create a ColorMatrix/Filter which would
eventually give us the drawable with the hue we need. This method works
with Android (setColorFilter API) instead of against it.

It looks pretty good on the darker themes (check out Google Blue) and
only a little off on some of the brighter ones.

Indeterminate progress indicators are not seen often in Hangouts. I
could only locate a few instances where it would appear consistently.
When it does show itself on Lollipop it's a green-blue color. This works
well for the default Google / Hangouts Green color not for other themes.
I used a variation  of the method mentioned above to achieve the
colorization. It's done at runtime as opposed to resource-load time. On
4.x the progress indicator is always gray so there's no need to style
it.

This seems relatively performant considering what it accomplishes but I
have yet to run any benchmarks.