Safely close the shell after it's done
This isn’t the ideal way to go about locking the screen but I tried the
Device Admin route and it was very difficult to do without opening up
the device to other security issues. The input binary seems to be a
little slow. It might be beneficial to compile my own…

As a side note the lockNow device admin method forces the device to
completely lock making it incompatible with Lollipop’s smart locking.