Add sound customization
Suggested by weaverinva @ XDA.

This module allows for users to swap out a given Hangouts sound with a
MP3 (or similar) on their device. A basic file chooser is provided for
pre-KitKat devices with no DocumentProvider interface.

Because the replacement method can be somewhat slow, it is recommend to
toggle this entire feature off if it is not being used or the device is
often on silent / vibrate.