Add the ability to hide the emoji button
The emoji button starts life as "gone" and depending on what a
particular method determines it is set to "visible" soon after creation
of the ComposeMessageView. This feature simply prevents that method
from executing.

This is incompatible with "attach anytime" since that feature requires
the availability of the emoji button.