Each module now responsible for disabling itself
Hooks will now be run even if XHangouts is “off” according to its own
configuration. This is so that less options will require Hangouts to
be restarted before use and allows for a more predictable assumption
of state. The downside is that XHangouts isn’t quite as disabled as
some users may seem to think it is. Luckily they have the Xposed
Installer to completely disable XHangouts when needed.

This behavior is inherently less performant but again, if XHangouts is
causing noticeable performance issues disabling it completely via the
Installer is the best course of action.

Note that this only applies to method hooks. Obtaining a valid Context
in handleInitPackageResources is hard™ so the modEnabled check is
still done in the XHangouts class for each module for resource hooks.