Refactored ImageResizing
Other than debug statements and a fix with PNG handling, there are no
functional changes

ImageResizing is very similar to MmsResizing and it might be
advantageous to move their common functionality into a utility class.

Rotation is done as the last step but it might need to be done first to
accurately reflect the XHangouts scale settings.