Fix the UiQuickSettings icon
For some reason the old way doesn’t work. Not sure what the problem
is. Could be a regression in Xposed or Hangouts is loading drawables
in a way that Xposed can’t handle. Who knows? What we’re doing now
is completely replacing a hopefully unused resource within Hangouts
and then setting that as our desired icon.

Since this is a little tricky to do when there is no context available
during handleLoadPackage we defer this hook until the Application
instance is available. Since we’re still hooking before the method is
called our replacement works.